+++ scripts/macros/2024/classFeatures/bard/lore/cuttingWords.js
@@ -16,7 +16,7 @@ async function damage({trigger, workflow}) {
     if (!nearbyTokens.length) return;
     for (let token of nearbyTokens) {
         let item = itemUtils.getItemByIdentifier(token.actor, 'cuttingWords');
-        let selection = await dialogUtils.confirm(item.name, genericUtils.format('CHRISPREMADES.Macros.CuttingWords.Damage', {item: item.name, name: token.document.name}, {userId: socketUtils.firstOwner(token.actor, true)}));
+        let selection = await dialogUtils.confirm(item.name, genericUtils.format('CHRISPREMADES.Macros.CuttingWords.Damage', {item: item.name, name: token.document.name}), {userId: socketUtils.firstOwner(token.actor, true)});
         if (!selection) continue;
         let result = await workflowUtils.syntheticItemRoll(item, [workflow.token], {consumeResources: true, userId: socketUtils.firstOwner(token.actor, true)});
         let total = -result.damageRolls[0].total;
@@ -68,7 +68,7 @@ async function attack({trigger, workflow}) {
     if (!nearbyTokens.length) return;
     for (let token of nearbyTokens) {
         let item = itemUtils.getItemByIdentifier(token.actor, 'cuttingWords');
-        let selection = await dialogUtils.confirm(item.name, genericUtils.format('CHRISPREMADES.Macros.CuttingWords.Attack', {item: item.name, name: token.document.name, attack: workflow.attackTotal}, {userId: socketUtils.firstOwner(token.actor, true)}));
+        let selection = await dialogUtils.confirm(item.name, genericUtils.format('CHRISPREMADES.Macros.CuttingWords.Attack', {item: item.name, name: token.document.name, attack: workflow.attackTotal}), {userId: socketUtils.firstOwner(token.actor, true)});
         if (!selection) continue;
         let result = await workflowUtils.syntheticItemRoll(item, [workflow.token], {consumeResources: true, userId: socketUtils.firstOwner(token.actor, true)});
         await workflowUtils.bonusAttack(workflow, String(-result.damageRolls[0].total));