+++ scripts/macros/2024/spells/beaconOfHope.js
@@ -1,6 +1,6 @@
 import {beaconOfHope as beaconOfHopeLegacy} from '../../2014/spells/beaconOfHope.js';
 export let beaconOfHope = {
     name: beaconOfHopeLegacy.name,
-    version: beaconOfHopeLegacy.version,
+    version: '1.2.32',
     midi: beaconOfHopeLegacy.midi
 };