+++ scripts/macros/2024/spells/shield.js
@@ -1,6 +1,6 @@
 import {shield as shieldLegacy} from '../../../legacyMacros.js';
 export let shield = {
     ...shieldLegacy,
-    version: '1.2.25',
+    version: '1.2.32',
     rules: 'modern'
 };